feat: consolidate exports into the Output panel (Output Center)
The Inspector owned three export sections, point-cloud file formats, the Visual
Export Studio image export, and the PDF report, while the Output panel owned the
converter, measurements, KML, and the integrity report. Two panels, both about
exporting.

Move the three Inspector sections into the Output panel so every deliverable
lives in one place. The UX is unchanged: the same collapsible sections, the same
buttons, the same load-time gating, only the mount point differs. The sections
and their gating are extracted into src/ui/export/exportDeliverables.ts; the
shared collapsibleSection helper moves to its own file so both panels use it.
The image-export and report gating is now driven on ExportPanel, threaded to the
open pipelines through the deps, and the point-format quick export hides while a
streaming scan is active. Inspector loses roughly 200 lines and its report and
export-type imports.

main.ts baseline moved 5824 to 5828 for the callback relocation and the deps
wiring; the architecture map and known-limitations counts follow.
