+++ docs/validation/monolith-size-baseline.json
@@ -1,7 +1,7 @@
 {
   "files": {
     "src/main.ts": {
-      "lines": 5824,
+      "lines": 5828,
       "goal": 2500
     },
     "src/render/Viewer.ts": {
